Create Curve.cpp

+++ Curve.cpp
@@ -115,7 +115,7 @@ const uint16_t Curve::etable[1024] PROGMEM = {
 };
 
 
-uint8_t Curve::exponential(uint16_t i) {
+uint16_t Curve::exponential(uint16_t i) {
  // If the compiler could build the above table for us, we could have "simply":
  //return !i ? 0 : round(exp(log(MAXOUTPUT) * i / MAXINPUT));
